Updated data types

+++ Queries/Schema.sql
@@ -4,10 +4,10 @@ CREATE TABLE spotify (
 	Artist_Name VARCHAR NOT NULL, 
 	Song_Name VARCHAR NOT NULL, 
 	Days INT NOT NULL, 
-	Top_10_xTimes INT NOT NULL, 
-	Peak_Position INT NOT NULL,
+	Top_10_xTimes FLOAT NOT NULL, 
+	Peak_Position FLOAT NOT NULL,
 	Peak_Position_xTimes VARCHAR NOT NULL,
-	Peak_Streams INT NOT NULL, 
-	Total_Streams INT NOT NULL,
+	Peak_Streams FLOAT NOT NULL, 
+	Total_Streams FLOAT NOT NULL,
 	PRIMARY KEY (Position)
 );
